Made routes more logical
