+++ views/layouts/layout.hbs
@@ -107,7 +107,7 @@
                         {{/ifCond}}
                         {{/unless}}
                         {{#if session.user}}
-							<li><a href="/logout"><i class="fa fa-sign-out" aria-hidden="true"> </i>Logout</a></li>
+							<li><a href="/admin/logout"><i class="fa fa-sign-out" aria-hidden="true"> </i>Logout</a></li>
 						{{/if}}
 					</ul>
 				</div>
